prost-types: Move tests to the respective modules.
