+++ prost-types/src/any.rs
@@ -49,3 +49,24 @@ impl Name for Any {
         type_url_for::<Self>()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn check_any_serialization() {
+        let message = Timestamp::date(2000, 1, 1).unwrap();
+        let any = Any::from_msg(&message).unwrap();
+        assert_eq!(
+            &any.type_url,
+            "type.googleapis.com/google.protobuf.Timestamp"
+        );
+
+        let message2 = any.to_msg::<Timestamp>().unwrap();
+        assert_eq!(message, message2);
+
+        // Wrong type URL
+        assert!(any.to_msg::<Duration>().is_err());
+    }
+}
